ci(insight.yml, package.json): updated report command, added publish option
