+++ package.json
@@ -47,7 +47,7 @@
     "release": "changeset publish",
     "release:dev": "changeset publish --tag dev",
     "release:next": "changeset publish --tag next",
-    "report:insight": "tsx scripts/report-insight",
+    "report:insights": "tsx scripts/report-insights",
     "notification:release": "tsx scripts/notification-release",
     "clean": "pnpm -r --parallel exec rimraf dist .turbo *.log node_modules",
     "docs:dev": "pnpm --prefix docs dev",
